requirements.txt
django
django-rest-framework
flake8
psycopg2
Pillow
python-dotenv
